+++ frontend/flask/Extensions.py
@@ -1,7 +1,10 @@
 from flask_sqlalchemy import SQLAlchemy
 from flask_caching import Cache
 from flask_login import LoginManager
+from authlib.integrations.flask_client import OAuth
 
 DATABASE = SQLAlchemy()
 CACHE= Cache()
 LOGIN_MANAGER= LoginManager()
+# Initialize Authlib OAuth client for Auth0
+oauth = OAuth()
